% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx_tx.R
\name{converge_aliases}
\alias{converge_aliases}
\title{Takes an input vector and replaces aliases with preferred values}
\usage{
converge_aliases(
  input_vector,
  input_vector_sep = "+",
  output_vector_sep = "+",
  lut_path,
  alias_clms = "Alias",
  preferred_name_clm = "Name",
  alias_sep = ",",
  include_missing_terms = FALSE,
  default_empty_value = NA
)
}
\arguments{
\item{input_vector}{Character vector with aliases to be replaced}

\item{input_vector_sep}{Character separating multi-name values of input vector}

\item{output_vector_sep}{Character to separate returned multi-name values}

\item{lut_path}{Path to .tsv with alias and preferred name columns}

\item{alias_clms}{Vector of column names containing aliases in lut}

\item{preferred_name_clm}{Name of column from which to pull preferred names to replace aliases}

\item{alias_sep}{Character separating multiple aliases found within a single alias_clm}

\item{include_missing_terms}{Boolean to return unchanged names that are not found in the alias_clms}

\item{default_empty_value}{Value to use where output vector name is empty or NA}
}
\value{
Returns a vector of same length as input_vector with aliases replaced by preferred values
}
\description{
This method replaces aliases in an input vector using a user-defined
lookup table.
}
\keyword{internal}
