% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_clm_case.R
\name{convert_clm_case}
\alias{convert_clm_case}
\title{Takes an input_string and puts it in title format with underscores.}
\usage{
convert_clm_case(input_string)
}
\arguments{
\item{input_string}{}
}
\value{
Returns dataframe with updated and added *_Rx and *_Tx columns
}
\description{
Takes an input_string and puts it in title format with underscores.
}
