% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depricated.R
\name{create_input_path}
\alias{create_input_path}
\title{Depricated method to create the full path to a given input file - set this manually with file.path instead}
\usage{
create_input_path(dataset, filename, subfolder = NA)
}
\arguments{
\item{dataset}{The name of the dataset ( which should match the folder name for this dataset in the filesystem )}

\item{filename}{The name of the input file}

\item{subfolder}{The path to the subfolder containing the input file, if any}
}
\value{
The full path to the file being requested starting from 
/datastore/nextgenout5/share/labs/Vincent_Lab/datasets
}
\description{
Uses the RAW_DATA_DIR and the parameters passed in to create a full path to the input file being requested
}
