% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_names_and_ids.R
\name{create_patient_ids}
\alias{create_patient_ids}
\title{Creates patient ids}
\usage{
create_patient_ids(dataset, patient_names)
}
\arguments{
\item{dataset}{Name of dataset to use with these ids}

\item{patient_names}{Vector of patient names to use in creation of ids}
}
\value{
Returns a vector of patient_ids
}
\description{
Concatenates dataset with patient_names.
}
