% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_names_and_ids.R
\name{create_run_ids}
\alias{create_run_ids}
\title{Creates run ids}
\usage{
create_run_ids(patient_ids, run_names)
}
\arguments{
\item{patient_ids}{Vector of patient_ids to use}

\item{run_names}{Vector of run_names to use}
}
\value{
Returns a vector of run_ids
}
\description{
Concatenates patient_ids with run_names.
}
