% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetprep.R
\docType{package}
\name{datasetprep-package}
\alias{datasetprep}
\alias{datasetprep-package}
\title{datasetprep: Basic Helpers for Dataset Prep}
\description{
Methods to make life easier when preparing datasets for analysis.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Benjamin-Vincent-Lab/datasetprep/issues}
}

}
\author{
\strong{Maintainer}: Nathan Wheeler \email{regenernate@gmail.com}

}
\keyword{internal}
