% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_somatic_workflow_matches.R
\name{find_somatic_workflow_matches}
\alias{find_somatic_workflow_matches}
\title{Sets field equal to patient name where data exists for somatic workflow}
\usage{
find_somatic_workflow_matches(dat)
}
\arguments{
\item{dat}{Data.frame from which to determine matches}
}
\value{
Vector of matches
}
\description{
Typically somatic workflow matches are when a given patient has samples of 
tumor data in RNA and DNA form along with normal tissue in DNA form. 
This method determines for which patients this is true.
}
\section{Limitations}{

\itemize{
  \item Requires that data.frame sent have columnes named Patient_Name, Normal,
  Analyte and Run_Name
}
}

