% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_methods.R
\name{get_all_columns}
\alias{get_all_columns}
\title{Returns character vector concatenation of get_patient_columns and get_run_only_columns}
\usage{
get_all_columns()
}
\value{
Character vector of get_patient_columns and get_run_only_columns
}
\description{
Method to return character vector combining of get_patient_columns and
get_run_only_columns
}
