% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_methods.R
\name{get_run_only_columns}
\alias{get_run_only_columns}
\title{Returns character vector of RUN_ONLY_COLUMNS}
\usage{
get_run_only_columns()
}
\value{
Character vector of RUN_ONLY_COLUMNS
}
\description{
Method to return static list of RUN_ONLY_COLUMNS
}
