% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_methods.R
\name{init_paths}
\alias{init_paths}
\title{Initializes output directory}
\usage{
init_paths(base_dir)
}
\arguments{
\item{base_dir}{The base directory in which the temp_data folder should be created for output}
}
\value{
the post processing directory path that was created
}
\description{
Sets up the post processing directory and prepares temp library folder if necessary
}
