% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_methods.R
\name{list_missing_columns}
\alias{list_missing_columns}
\title{Prints any column names not yet defined in data}
\usage{
list_missing_columns(dat, col.names = NA)
}
\description{
Method to print out any col.names not yet defined in the data passed in. Intended to help with preparing valid data.
}
