% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx_tx.R
\name{lookup_drug_properties}
\alias{lookup_drug_properties}
\title{Looks up properties for drug combinations in input vector using user-defined lookup table}
\usage{
lookup_drug_properties(
  input_vector,
  property_lut_path = system.file("rx_table", "rx_table.tsv", package = "datasetprep"),
  copy_clms = c("ICI_Pathway", "ICI_Target"),
  all_properties = c("ICI", "Non_ICI", "IS", "aPD1", "aCTLA4", "aVEGF", "aBRAF", "aMAPK",
    "Chemo", "Steroid"),
  return_input = FALSE
)
}
\arguments{
\item{input_vector}{Character vector with drug names to be looked up}

\item{property_lut_path}{Path to .tsv with drug names and properties}

\item{copy_clms}{Optional vector of columns containing values to be copied directly}

\item{all_properties}{Vector of property names to look for in property_clm and include in return df}

\item{return_input}{Boolean whether to include the input vector as the first column of the return df}
}
\value{
Returns a dataframe with original input vector and itemized, 
boolean and copied properties as requested through input parameters.
}
\description{
This method calls base lookup_properties method with appropriate values for package drug table
and returns the resulting df with aCTLA4_PD1_Tx column added
}
