% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx_tx.R
\name{lookup_properties}
\alias{lookup_properties}
\title{Finds properties for names in input vector using user-defined lookup 
table}
\usage{
lookup_properties(
  input_vector,
  property_lut_path,
  name_clm = "Name",
  property_clm = "Properties",
  copy_clms = NA,
  all_properties = c(),
  input_vector_sep = " + ",
  property_sep = ",",
  output_property_sep = " + ",
  output_copy_clm_sep = " + ",
  itemized_clm_suffix = "_Itemized",
  boolean_clm_suffix = "_Boolean",
  no_info_itemized_value = NA_character_,
  no_info_boolean_value = FALSE,
  no_info_copy_value = NA,
  skip_itemized_clms = FALSE,
  skip_boolean_clms = FALSE,
  return_input = TRUE
)
}
\arguments{
\item{input_vector}{Character vector with names to be looked up}

\item{property_lut_path}{Path to .tsv with name and property columns}

\item{name_clm}{Column containing names in lut}

\item{property_clm}{Column containing properties in lut}

\item{copy_clms}{Some columns in the lookup table can have values that can be copied over.  For example, For 'Pembro' we'd want to find pembro on the list and then grab the info for ICI_Pathway (PD1) and pass that info along to the sample data. Optional vector of columns containing values to be copied directly}

\item{all_properties}{Vector of property names to look for in property_clm and include in return df}

\item{input_vector_sep}{Character separating multi-name values of input vector}

\item{property_sep}{Character separating multi-property values in property_clm}

\item{output_property_sep}{Character to separate multi-name values returned in itemized columns}

\item{output_copy_clm_sep}{Character separating copy column values from multi-name inputs}

\item{itemized_clm_suffix}{Characters appended to each property to form itemized column names in return df}

\item{boolean_clm_suffix}{Characters appended to each property to form boolean column names in return df}

\item{no_info_itemized_value}{Value to return for itemized columns where the given property is not found}

\item{no_info_boolean_value}{Value to return for boolean columns where the given property is not found}

\item{no_info_copy_value}{Value to return for copied columns where value is empty/NA}

\item{skip_itemized_clms}{Boolean to include itemized columns in return df}

\item{skip_boolean_clms}{Boolean to include boolean columns in return df}

\item{return_input}{Boolean whether to include the input vector as the first column of the return df}
}
\value{
Returns a dataframe with original input vector and itemized, 
boolean and copied values as requested through input parameters.
}
\description{
This method looks up names in user-defined lookup table, takes properties 
found in the associated property_clm and expands them into a return 
data.frame with itemized and boolean columns for each.
}
\keyword{internal}
