% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_methods.R
\name{output_summary}
\alias{output_summary}
\title{Outputs summary of values}
\usage{
output_summary(vec, varname)
}
\arguments{
\item{vec}{Vector to summarize}

\item{varname}{Name of values}
}
\value{
outputs values
}
\description{
Outputs summary of values
}
