% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_variable_summary.R
\name{output_variable_summary}
\alias{output_variable_summary}
\title{Outputs summary of variables in data.frame}
\usage{
output_variable_summary(
  dat = NULL,
  column_names = NULL,
  max_length = 10,
  title = NULL
)
}
\arguments{
\item{dat}{Data.frame from which to output summary}

\item{column_names}{Vector of column names to summarize. Defaults to all colnames}

\item{title}{Optional string to display as label for data being output}

\item{min_length}{Integer of maximum unique values to display for a given column}
}
\description{
In order to confirm data being generated through the prep phase, 
it can be helpful to output a summary of variables in a data.frame. 
This method takes a dataframe and does just that, outputting the summary to the console.
}
