% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_methods.R
\name{set_response_data}
\alias{set_response_data}
\title{Formats response data and sets associated fields}
\usage{
set_response_data(dat, response_col = "Response")
}
\arguments{
\item{dat}{Dataframe containing response values}

\item{response_col}{Name of column containing Response values}
}
\value{
Returns dat with updated and added columns
}
\description{
This method standardizes the Best Response data to full names in CamelCase 
and then sets fields that are dependent on those values ( Progression, 
Clinical_Benefit and Responder ).
}
\section{Limitations}{

\itemize{
  \item Only works for known set of responses ( PD, SD, PR, CR )
}
}

