% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rx_tx.R
\name{set_rx_tx}
\alias{set_rx_tx}
\title{Takes a drug list and returns fomatted variables}
\usage{
set_rx_tx(rx_list)
}
\arguments{
\item{rx_list}{Dataframe containing " + " separated Drugs}
}
\value{
Returns rx_list with updated and added columns
}
\description{
This method looks up drugs by name, normalizes them and sets the various 
pathway fields ( i.e. ICI_Rx, ICI_Tx, aPD1_Rx, aPD1_Tx ).
}
\section{Limitations}{

\itemize{
  \item Does not work for prior or subsq fields.
  \item Does not set NeoICI_Tx
  \item Does not handle non-Rx treatments like Surgery, Radiation, etc.
}
}

