% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_methods.R
\name{configure_output_columns}
\alias{configure_output_columns}
\title{Defines columns for output data}
\usage{
configure_output_columns()
}
\description{
Method to set Run and Patient columns to be outputted in final data.
}
