% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_methods.R
\name{configure_output_columns}
\alias{configure_output_columns}
\title{Depricated method to define columns for output data - use get_run_only_columns, get_patient_only_columns, get_all_columns instead}
\usage{
configure_output_columns()
}
\description{
Method to set RUN_ONLY_COLUMNS and PATIENT_ONLY_COLUMNS to be output to final data.
}
