% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_methods.R
\name{configure_readme}
\alias{configure_readme}
\title{Configure readme file}
\usage{
configure_readme(output_dir, dataset)
}
\arguments{
\item{output_dir}{path in which to save readme file}

\item{dataset}{name of dataset to use in readme file}
}
\description{
Method sets the readme path and removes existing readme file
}
