% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rx_tx.R
\name{converge_drug_aliases}
\alias{converge_drug_aliases}
\title{Takes an input vector of drug names/combinations and replaces 
aliases with preferred names}
\usage{
converge_drug_aliases(
  input_vector,
  drug_lut_path = system.file("rx_table", "rx_table.tsv", package = "datasetprep"),
  include_missing_terms = FALSE,
  default_empty_value = NA
)
}
\arguments{
\item{input_vector}{Character vector with drug aliases to be replaced}

\item{drug_lut_path}{Path to .tsv drug table}

\item{include_missing_terms}{Boolean to keep unchanged names that are not found in the alias_clms}

\item{default_empty_value}{Value to use where output vector name is empty or NA}
}
\value{
Returns a vector of same length as input_vector with drug aliases replaced by preferred names
}
\description{
This method replaces aliases in an input vector using a user-defined
drug table.
}
