% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_methods.R
\name{create_input_path}
\alias{create_input_path}
\title{Creates the full path to a given input file}
\usage{
create_input_path(dataset, filename)
}
\arguments{
\item{dataset}{The name of the dataset ( which should match the folder name for this dataset in the filesystem )}

\item{filename}{The name of the input file}
}
\value{
The full path to the file being requested
}
\description{
Uses the RAW_DATA_DIR and the parameters passed in to create a full path to the input file being requested
}
