% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient_names_and_ids.R
\name{format_patient_names}
\alias{format_patient_names}
\title{Formats patient names}
\usage{
format_patient_names(
  old_names,
  reduce_size = FALSE,
  min_chars = 1,
  prefix = "p"
)
}
\arguments{
\item{old_names}{Vector of patient names to format}

\item{reduce_size}{Boolean to indicate whether numeric portion of ids should be reduced to shortest length representing unique values}

\item{min_chars}{Integer minimum number of characters to keep in numeric portion of id's ( final lengths will be length(prefix)+min_char because prefix is prepended to numeric portion )}

\item{prefix}{Character value to prepend onto numeric portions of reformatted ids - defaults to "p"}
}
\value{
Returns the updated patient names vector
}
\description{
Data's original patient identifiers are often more complex than we want. 
This method takes those names ( within a limited subset of formats ) 
and reformats them as "p001" where the number portion is unique across patients.
}
\section{Limitations}{

\itemize{
  \item Only works properly if old names have unique numerical identifiers in them
  i.e. Patient1, Patient2, etc.
  \item Only works if old names have a single numerical element within them
  i.e wouldn't work with identifiers like MD-27-AR-0215
}
}

