% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_methods.R
\name{format_sex}
\alias{format_sex}
\title{Formats sex values}
\usage{
format_sex(
  sex,
  female_values = c("female", "f", "w", "woman", "girl", "g"),
  male_values = c("male", "m", "man", "boy", "b"),
  female_replacement = "Female",
  male_replacement = "Male"
)
}
\arguments{
\item{sex}{Vector of sex values}

\item{female_values}{String values in sex to replace with Female, defaults include "female", "F", and "f"}

\item{male_values}{String values in sex to replace with Male, defaults include "male", "M", and "m"}

\item{female_replacement}{Value to replace female_chars with, defaults to "Female"}

\item{male_replacement}{Value to replace male_chars with, defaults to "Male"}
}
\value{
Returns the updated sex vector
}
\description{
Converts data's original sex characters to standard Female / Male.
}
