% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_methods.R
\name{format_sex}
\alias{format_sex}
\title{Formats sex values}
\usage{
format_sex(sex, female_char = "F", male_char = "M")
}
\arguments{
\item{sex}{Vector of sex values}

\item{female_char}{Character used in sex to represent Female, defaults to F}

\item{male_char}{Character used in sex to represent Male, defaults to M}
}
\value{
Returns the updated sex vector
}
\description{
Converts data's original sex characters to standard Female / Male.
}
