% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_methods.R
\name{get_all_columns}
\alias{get_all_columns}
\title{Returns character vector of all output columns}
\usage{
get_all_columns()
}
\value{
Character vector of RUN_COLUMNS
}
\description{
Method to return static list of RUN_COLUMNS which is 
concatenation of RUN_ONLY_COLUMNS and PATIENT_ONLY_COLUMNS
}
