% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_methods.R
\name{get_patient_columns}
\alias{get_patient_columns}
\title{Returns character vector of PATIENT_ONLY_COLUMNS}
\usage{
get_patient_columns()
}
\value{
Character vector of PATIENT_ONLY_COLUMNS
}
\description{
Method to return static list of PATIENT_ONLY_COLUMNS
}
