% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configuration_methods.R
\name{get_run_only_columns}
\alias{get_run_only_columns}
\title{Returns character vector of column names for data only relevant to a 
sample or sequencing run as opposed to a patient.}
\usage{
get_run_only_columns()
}
\value{
Character vector of column names
}
\description{
Returns character vector of column names for data only relevant to a 
sample or sequencing run as opposed to a patient.
}
