% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rx_tx.R
\name{init_prior_and_subsq_rx_tx}
\alias{init_prior_and_subsq_rx_tx}
\title{Initializes prior and subsq Rx and Tx fields}
\usage{
init_prior_and_subsq_rx_tx(dat, prior_is_known = FALSE, subsq_is_known = FALSE)
}
\arguments{
\item{dat}{Dataframe to add columns to}

\item{prior_is_known}{Whether we want to initialize prior values as 
unknown ("Unknown", NA) or known ("None", FALSE). Defaults to FALSE ( unkonwn )}

\item{subsq_is_known}{Whether we want to initialize subsq values as 
unknown ("Unknown", NA) or known ("None", FALSE). Defaults to FALSE ( unkonwn )}
}
\value{
Returns dataframe with columns
}
\description{
This method sets the initial values for the various Rx and Tx fields for 
Prior and Subsq treatments
}
