% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatting_methods.R
\name{set_race_fields}
\alias{set_race_fields}
\title{Formats race values and sets individual race fields}
\usage{
set_race_fields(dat, input_clm = "Race", na_to_lc = FALSE)
}
\arguments{
\item{dat}{Dataframe containing race values}

\item{input_clm}{Name of column containing race values, defaults to Race, if NA intializes Race as all NA}

\item{na_to_lc}{Boolean dictating whether to set any na's to Likely Caucasian}
}
\value{
Returns dat with updated and added columns
}
\description{
Takes data's original race values and cleans them up for consistency, 
converts unknown races to Other, 
and sets the individual boolean race fields ( Caucasian, Asian, etc. ) based on updated Race.
}
