% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_rx_tx.R
\name{set_rx_tx}
\alias{set_rx_tx}
\title{Takes a drug list and returns formatted variables}
\usage{
set_rx_tx(
  dat,
  drug_table_path = system.file("rx_table", "rx_table.tsv", package = "datasetprep")
)
}
\arguments{
\item{dat}{Dataframe containing " + " separated Drugs as Drug variable}

\item{drug_table_path}{Full path to the drug table to be used}
}
\value{
Returns dataframe with updated and added *_Rx and *_Tx columns
}
\description{
This method looks up drugs by name, normalizes them and sets the various 
pathway fields ( i.e. ICI_Rx, ICI_Tx, aPD1_Rx, aPD1_Tx ).
}
\section{Limitations}{

\itemize{
  \item Does not work for prior or subsq fields.
  \item Does not set NeoICI_Tx
  \item Does not handle non-Rx treatments like Surgery, Radiation, etc.
}
}

